# new_server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "new_server API",
    "version": "1.0.0",
    "description": "Machine-readable API contract for the Express backend in new_server."
  },
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key"
      },
      "apiKeyAuthScheme": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Alternative format: `Authorization: ApiKey <apiKey>`"
      },
      "inboundBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Bearer token for worker ingress"
      }
    },
    "schemas": {
      "RequestId": {
        "type": "string"
      },
      "ErrorResponse": {
        "type": "object",
        "required": ["error", "requestId"],
        "properties": {
          "error": { "type": "string" },
          "details": {},
          "requestId": { "$ref": "#/components/schemas/RequestId" }
        }
      },
...
      "GrantedBy": {
        "oneOf": [
          {
            "type": "object",
            "required": ["userId", "username", "displayName"],
            "properties": {
              "userId": { "type": "integer" },
              "username": { "type": "string" },
              "displayName": { "type": "string", "nullable": true }
            }
          },
          {
            "type": "object",
            "required": ["label"],
            "properties": {
              "label": { "type": "string" }
            }
          },
          {
            "type": "null"
          }
        ]
      },
      "PermissionUser": {
        "type": "object",
        "required": ["id", "username", "displayName", "telegramId", "status"],
        "properties": {
          "id": { "type": "integer" },
          "username": { "type": "string" },
          "displayName": { "type": "string" },
          "telegramId": { "type": "string", "nullable": true },
          "status": { "type": "string", "enum": ["active", "disabled"] }
        }
      },
      "Permission": {
        "type": "object",
        "required": ["id", "domain", "localPart", "emailAddress", "role", "status", "user", "grantedBy", "createdAt", "updatedAt"],
        "properties": {
          "id": { "type": "integer" },
          "domain": { "type": "string" },
          "localPart": { "type": "string", "nullable": true },
          "emailAddress": { "type": "string", "nullable": true },
          "role": { "type": "string", "enum": ["viewer", "operator", "admin"] },
          "status": { "type": "string", "enum": ["active", "disabled"] },
          "user": { "$ref": "#/components/schemas/PermissionUser" },
          "grantedBy": { "$ref": "#/components/schemas/GrantedBy" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "User": {
        "type": "object",
        "required": ["id", "username", "displayName", "telegramId", "status", "hasPassword", "hasApiKey", "apiKeyLastFour", "isAdmin", "lastSeenAt", "createdAt", "updatedAt"],
        "properties": {
          "id": { "type": "integer" },
          "username": { "type": "string" },
          "displayName": { "type": "string" },
          "telegramId": { "type": "string", "nullable": true },
          "status": { "type": "string", "enum": ["active", "disabled"] },
          "hasPassword": { "type": "boolean" },
          "hasApiKey": { "type": "boolean" },
          "apiKeyLastFour": { "type": "string", "nullable": true },
          "isAdmin": { "type": "boolean" },
          "lastSeenAt": { "type": "string", "format": "date-time", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "UserWithPermissions": {
        "allOf": [
          { "$ref": "#/components/schemas/User" },
          {
            "type": "object",
            "required": ["permissions"],
            "properties": {
              "permissions": {
                "type": "array",
                "items": { "$ref": "#/components/schemas/Permission" }
              }
            }
          }
        ]
      },
...
      "DomainCounts": {
        "type": "object",
        "required": ["domainPermissions", "mailboxPermissions", "emails"],
        "properties": {
          "domainPermissions": { "type": "integer" },
          "mailboxPermissions": { "type": "integer" },
          "emails": { "type": "integer" }
        }
      },
      "Domain": {
        "type": "object",
        "required": ["id", "domain", "description", "status", "inboundEnabled", "isDefault", "counts", "createdAt", "updatedAt"],
        "properties": {
          "id": { "type": "integer" },
          "domain": { "type": "string" },
          "description": { "type": "string" },
          "status": { "type": "string", "enum": ["active", "disabled"] },
          "inboundEnabled": { "type": "boolean" },
          "isDefault": { "type": "boolean" },
          "counts": { "$ref": "#/components/schemas/DomainCounts" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "AdminMutationRequest": {
        "type": "object",
        "properties": {
          "userId": { "oneOf": [{ "type": "integer" }, { "type": "string" }] },
          "username": { "type": "string" }
        }
      },
      "PermissionCreateRequest": {
        "type": "object",
        "required": ["role"],
        "properties": {
          "userId": { "oneOf": [{ "type": "integer" }, { "type": "string" }] },
          "username": { "type": "string" },
          "displayName": { "type": "string" },
          "telegramId": { "type": "string", "nullable": true },
          "domain": { "type": "string" },
          "localPart": { "type": "string", "nullable": true },
          "role": { "type": "string", "enum": ["viewer", "operator", "admin"] },
          "status": { "type": "string", "enum": ["active", "disabled"] }
        }
      },
      "PermissionUpdateRequest": {
        "type": "object",
        "properties": {
          "role": { "type": "string", "enum": ["viewer", "operator", "admin"] },
          "status": { "type": "string", "enum": ["active", "disabled"] }
        }
      },
      "DomainUpsertRequest": {
        "type": "object",
        "required": ["domain"],
        "properties": {
          "domain": { "type": "string" },
          "description": { "type": "string" },
          "status": { "type": "string", "enum": ["active", "disabled"] },
          "inboundEnabled": { "type": "boolean" },
          "isDefault": { "type": "boolean" }
        }
      },
      "DomainUpdateRequest": {
        "type": "object",
        "properties": {
          "description": { "type": "string" },
          "status": { "type": "string", "enum": ["active", "disabled"] },
          "inboundEnabled": { "type": "boolean" },
          "isDefault": { "type": "boolean" }
        }
      },
...
      "SimpleSuccessResponse": {
        "type": "object",
        "required": ["success", "requestId"],
        "properties": {
          "success": { "type": "boolean" },
          "requestId": { "$ref": "#/components/schemas/RequestId" }
        }
      }
    },
    "responses": {
...
      "NotFound": {
        "description": "Not found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "Conflict": {
        "description": "Conflict",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      }
    }
  },
  "paths": {
...
    "/v1/admins": {
      "get": {
        "tags": ["Admins"],
        "summary": "List admins",
        "description": "Global admin only",
        "security": [{ "bearerAuth": [] }, { "apiKeyHeader": [] }, { "apiKeyAuthScheme": [] }],
        "responses": {
          "200": {
            "description": "Admin list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "count": { "type": "integer" },
                    "admins": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/User" }
                    },
                    "requestId": { "$ref": "#/components/schemas/RequestId" }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Admins"],
        "summary": "Grant admin role",
        "description": "Global admin only",
        "security": [{ "bearerAuth": [] }, { "apiKeyHeader": [] }, { "apiKeyAuthScheme": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AdminMutationRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Admin granted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "admin": { "$ref": "#/components/schemas/UserWithPermissions" },
                    "requestId": { "$ref": "#/components/schemas/RequestId" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/admins/{userId}": {
      "delete": {
        "tags": ["Admins"],
        "summary": "Revoke admin role",
        "description": "Global admin only",
        "security": [{ "bearerAuth": [] }, { "apiKeyHeader": [] }, { "apiKeyAuthScheme": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Admin revoked",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SimpleSuccessResponse" }
              }
            }
          },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
...
    "/v1/domains": {
      "get": {
        "tags": ["Domains"],
        "summary": "List domains",
        "security": [{ "bearerAuth": [] }, { "apiKeyHeader": [] }, { "apiKeyAuthScheme": [] }],
        "responses": {
          "200": {
            "description": "Domain list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "count": { "type": "integer" },
                    "domains": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Domain" }
                    },
                    "requestId": { "$ref": "#/components/schemas/RequestId" }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Domains"],
        "summary": "Create or upsert domain",
        "description": "Global admin only",
        "security": [{ "bearerAuth": [] }, { "apiKeyHeader": [] }, { "apiKeyAuthScheme": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DomainUpsertRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Domain created or updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "domain": { "$ref": "#/components/schemas/Domain" },
                    "requestId": { "$ref": "#/components/schemas/RequestId" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/domains/{domain}": {
      "get": {
        "tags": ["Domains"],
        "summary": "Get domain",
        "security": [{ "bearerAuth": [] }, { "apiKeyHeader": [] }, { "apiKeyAuthScheme": [] }],
        "parameters": [
          { "in": "path", "name": "domain", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Domain details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "domain": { "$ref": "#/components/schemas/Domain" },
                    "requestId": { "$ref": "#/components/schemas/RequestId" }
                  }
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["Domains"],
        "summary": "Update domain",
        "description": "Requires domain admin or global admin",
        "security": [{ "bearerAuth": [] }, { "apiKeyHeader": [] }, { "apiKeyAuthScheme": [] }],
        "parameters": [
          { "in": "path", "name": "domain", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DomainUpdateRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Domain updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "domain": { "$ref": "#/components/schemas/Domain" },
                    "requestId": { "$ref": "#/components/schemas/RequestId" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/domains/{domain}/permissions": {
      "get": {
        "tags": ["Domains"],
        "summary": "List permissions in a domain",
        "description": "Requires domain admin or global admin",
        "security": [{ "bearerAuth": [] }, { "apiKeyHeader": [] }, { "apiKeyAuthScheme": [] }],
        "parameters": [
          { "in": "path", "name": "domain", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Domain permission list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "count": { "type": "integer" },
                    "permissions": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Permission" }
                    },
                    "requestId": { "$ref": "#/components/schemas/RequestId" }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Domains"],
        "summary": "Create or upsert domain permission",
        "description": "Requires domain admin or global admin",
        "security": [{ "bearerAuth": [] }, { "apiKeyHeader": [] }, { "apiKeyAuthScheme": [] }],
        "parameters": [
          { "in": "path", "name": "domain", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  { "$ref": "#/components/schemas/PermissionCreateRequest" },
                  {
                    "type": "object",
                    "properties": {
                      "domain": { "readOnly": true }
                    }
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Permission created or updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "permission": { "$ref": "#/components/schemas/Permission" },
                    "requestId": { "$ref": "#/components/schemas/RequestId" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/domains/{domain}/permissions/{permissionId}": {
      "get": {
        "tags": ["Domains"],
        "summary": "Get permission in domain",
        "description": "Requires domain admin or global admin",
        "security": [{ "bearerAuth": [] }, { "apiKeyHeader": [] }, { "apiKeyAuthScheme": [] }],
        "parameters": [
          { "in": "path", "name": "domain", "required": true, "schema": { "type": "string" } },
          { "in": "path", "name": "permissionId", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Permission details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "permission": { "$ref": "#/components/schemas/Permission" },
                    "requestId": { "$ref": "#/components/schemas/RequestId" }
                  }
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "patch": {
        "tags": ["Domains"],
        "summary": "Update permission in domain",
        "description": "Requires domain admin or global admin",
        "security": [{ "bearerAuth": [] }, { "apiKeyHeader": [] }, { "apiKeyAuthScheme": [] }],
        "parameters": [
          { "in": "path", "name": "domain", "required": true, "schema": { "type": "string" } },
          { "in": "path", "name": "permissionId", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PermissionUpdateRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Permission updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "permission": { "$ref": "#/components/schemas/Permission" },
                    "requestId": { "$ref": "#/components/schemas/RequestId" }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Domains"],
        "summary": "Delete permission in domain",
        "description": "Requires domain admin or global admin",
        "security": [{ "bearerAuth": [] }, { "apiKeyHeader": [] }, { "apiKeyAuthScheme": [] }],
        "parameters": [
          { "in": "path", "name": "domain", "required": true, "schema": { "type": "string" } },
          { "in": "path", "name": "permissionId", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Permission deleted",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SimpleSuccessResponse" }
              }
            }
          }
        }
      }
    },
...
  }
}